Revert "Move Victory-specific classes into seperate package"
This reverts commit 3fa49a99b859418323995eae88c0832e2bfc147d.
