+++ org.emoflon.ibex.tgg.core.runtime/src/org/emoflon/ibex/tgg/operational/monitoring/IVictoryDataProvider.java
@@ -1,4 +1,4 @@
-package org.emoflon.ibex.tgg.operational.victory;
+package org.emoflon.ibex.tgg.operational.monitoring;
 
 import java.io.IOException;
 import java.util.Collection;
